Update styles.css
Assesment requirements completed.

To display the navigation items in a row, with an equal amount of space between the items.
The navigation bar should scroll with the content and then become stuck at the top of the viewport when it reaches it.
The image that is inside the article should have text wrapped around it.
The <article> and <aside> elements should display as a two column layout. The columns should be a flexible size so that if the browser window shrinks smaller the columns become narrower.
The photographs should display as a two column grid with a 1 pixel gap between the images.